+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Computed fields now allow the computing function to be passed as a string containing the name of a method, rather than the function object itself.
 - `ListField` and `SetField` can now still be pickled when a non-callable default is specified. This was preventing them being used
  in migrations.
+- Improve the approx SQL representation of Datastore commands (update, delete etc.)
 
 ### Bug fixes: